this commit first introduced a start of API, though it's still using the same TODO: ask to artyom how to use a skin for xml stuffs and a skin for normal web view  * also fix a lot of warnings when compiling with -pendantic and -Wall  * put a work around , the translations vectors has one two extra empty rows which are unneeded and have been reveleaed thanks to the api  * started to use reference in view to avoid useless copies (thanks to qdii advice)

+++ src/models/SearchEngine.cpp
@@ -348,7 +348,7 @@ void SearchEngine::remove_meta(
 results::Searches SearchEngine::search(
     std::string request,
     std::string lang,
-    int size,
+    unsigned int size,
     int offset
 ) {
     std::set<int> uniqueResults;
@@ -391,7 +391,7 @@ results::Searches SearchEngine::search(
 results::Searches SearchEngine::search_one_index(
     std::string request,
     std::string indexName,
-    int size,
+    unsigned int size,
     int offset,
     results::Searches& results,
     std::set<int>& uniqueResults
@@ -413,7 +413,7 @@ results::Searches SearchEngine::search_one_index(
         &rnum
     );
 
-    int max = std::min(rnum,size+offset);
+    int max = std::min(rnum,(int)size+offset);
 
     if (beginWithResult) {
